+++ src/rules/no-restricted-paths.js
@@ -32,6 +32,7 @@ module.exports = {
                   },
                   uniqueItems: true,
                 },
+                message: { type: 'string' },
               },
               additionalProperties: false,
             },
@@ -102,7 +103,7 @@ module.exports = {
 
           context.report({
             node,
-            message: `Unexpected path "{{importPath}}" imported in restricted zone.`,
+            message: `Unexpected path "{{importPath}}" imported in restricted zone.${zone.message ? ` ${zone.message}` : ''}`,
             data: { importPath },
           })
         })